+++ src/app/app.module.ts
@@ -6,6 +6,8 @@ import { SectionComponent } from './components/section/section.component';
 import { PreviewComponent } from './components/preview/preview.component';
 import { MetaDataComponent } from './components/meta-data/meta-data.component';
 import { MainComponent } from './components/main/main.component';
+import { kEntrySectionsService } from './services/k-entry-sections.service';
+
 
 @NgModule({
   declarations: [
@@ -18,7 +20,7 @@ import { MainComponent } from './components/main/main.component';
   imports: [
     BrowserModule
   ],
-  providers: [],
+  providers: [kEntrySectionsService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
