+++ src/app/components/preview/preview.component.html
@@ -1,4 +1,4 @@
-<div class="priview">
+<div class="preview">
   <div class="item">
     <div class="player">
       <div class="upload">
@@ -6,9 +6,9 @@
       </div>
     </div>
     <div class="player ">
-      <div class="priviewItem">
+      <div class="previewItem">
         <span class="kIconshare share"></span>
-        <span> priview & Embed</span>
+        <span> preview & Embed</span>
       </div>
     </div>
   </div>
