+++ src/app/components/preview/preview.component.scss
@@ -1,6 +1,6 @@
 @import '../../../theme/_colors.scss';
 @import '../../../theme/_mixins.scss';
-.priview {
+.preview {
   display: flex;
   flex-direction: column;
   flex: 0 0 220px; //padding: 10px;
@@ -16,7 +16,7 @@
       display: flex;
       flex-direction: column;
       height: 100%;
-      .priviewItem {
+      .ppreviewItem {
         padding: 10px;
         display: flex;
         flex: 0 0 auto;
